+++ bc_obps/registration/schema/v1/__init__.py
@@ -1,6 +1,4 @@
 # ruff: noqa
-from .operation import *
-from .operator import *
 from .user_operator import *
 from .user import *
 from .regulated_products import *
